requirements.txt
fastapi
uvicorn
python-multipart
motor
pydantic
PyMuPDF
pinecone
langchain
langchain-google-genai
langchain-pinecone
python-dotenv
